% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-multichannel.R
\name{plotPSD}
\alias{plotPSD}
\title{Plot power spectral density}
\usage{
plotPSD(x, channels = NULL, sample = 1L, log_scale = TRUE, freq_range = NULL)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{channels}{Integer vector of channel indices. If NULL, plots all.}

\item{sample}{Integer index for the sample (for 3D data).}

\item{log_scale}{Logical. If TRUE, uses log scale for power.}

\item{freq_range}{Numeric vector of length 2 specifying frequency range.}
}
\value{
A ggplot object.
}
\description{
Generates a PSD plot for specified channels.
}
\references{
Wickham, H. (2016). \emph{ggplot2: Elegant Graphics for Data
Analysis}. Springer-Verlag New York. \doi{10.1007/978-3-319-24277-4}

Delorme, A. & Makeig, S. (2004). "EEGLAB: an open source toolbox for
analysis of single-trial EEG dynamics including independent component
analysis." \emph{Journal of Neuroscience Methods}, 134(1), 9-21.
\doi{10.1016/j.jneumeth.2003.10.009}
}
\seealso{
\code{\link[=plotSignal]{plotSignal()}} for time-domain signal plots, \code{\link[=fftSignals]{fftSignals()}} for
FFT computation, \code{\link[=bandPower]{bandPower()}} for frequency band power extraction,
\code{\link[=plotSpectrogram]{plotSpectrogram()}} for time-frequency visualization.
}
