% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-stats.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{icc}
\alias{sem}
\alias{mdc}
\alias{blandAltman}
\alias{cohensD}
\alias{etaSquared}
\alias{waveformCMC}
\alias{waveformICC}
\alias{waveformReliability}
\alias{fPCA}
\alias{reconstructFPCA}
\alias{registerCurves}
\alias{circularSummary}
\alias{rayleighTest}
\alias{watsonWilliamsTest}
\alias{circularLinearCorrelation}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{PhysioExperiment}{\code{\link[PhysioExperiment]{blandAltman}}, \code{\link[PhysioExperiment]{circularLinearCorrelation}}, \code{\link[PhysioExperiment]{circularSummary}}, \code{\link[PhysioExperiment]{cohensD}}, \code{\link[PhysioExperiment]{etaSquared}}, \code{\link[PhysioExperiment]{fPCA}}, \code{\link[PhysioExperiment]{icc}}, \code{\link[PhysioExperiment]{mdc}}, \code{\link[PhysioExperiment]{rayleighTest}}, \code{\link[PhysioExperiment]{reconstructFPCA}}, \code{\link[PhysioExperiment]{registerCurves}}, \code{\link[PhysioExperiment]{sem}}, \code{\link[PhysioExperiment]{watsonWilliamsTest}}, \code{\link[PhysioExperiment]{waveformCMC}}, \code{\link[PhysioExperiment]{waveformICC}}, \code{\link[PhysioExperiment]{waveformReliability}}}
}}

